Move PrefixEnvironmentVariable
